+++ packages/core/src/index.js
@@ -2,9 +2,7 @@ module.exports = {
   evaluation: require('./code-evaluation'),
   io: require('./io'),
   loading: require('./code-loading'),
-  observableUtils: require('./observable-utils'),
   parameters: require('./parameters'),
-  sideEffects: require('./sideEffects'),
   utils: require('./utils'),
   web: require('./web')
 }
