--- packages/core/src/sideEffects/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  dom: require('./dom'),
-  i18n: require('./i18n'),
-  titleBar: require('./titleBar'),
-  worker: require('./worker')
-}
