+++ packages/web/package.json
@@ -42,11 +42,16 @@
     "@jscad/vtree": "2.0.3",
     "brace": "0.11.1",
     "codemirror": "5.57.0",
+    "es2015-i18n-tag": "1.6.1",
     "file-saver": "2.0.2",
+    "is-electron": "2.2.0",
+    "morphdom": "2.6.1",
     "most": "1.8.0",
+    "@most/create": "2.0.1",
     "most-gestures": "0.4.2",
     "most-proxy": "3.3.0",
-    "nanohtml": "1.9.1"
+    "nanohtml": "1.9.1",
+    "webworkify": "1.5.0"
   },
   "devDependencies": {
     "ava": "3.15.0",
