+++ packages/web/src/ui/flow/tools.js
@@ -1,6 +1,6 @@
 const most = require('most')
 
-const { withLatestFrom } = require('@jscad/core').observableUtils
+const { withLatestFrom } = require('../../most-utils')
 
 const actions = ({ sources }) => {
   const setActiveTool$ = most.mergeArray([
